+++ app/javascript/packs/resultPage/ManualResultPage.jsx
@@ -258,7 +258,7 @@ const ManualResultPage = ({award, mastery, instructorId, boys, customDescription
         <Page>
           <View>
             <Header>              
-              <Image style={styles.logo} src="/packs/media/packs/general/bb-crest-7106b85f04ce6829d39a973203d05a81.png"></Image>
+              <Image style={styles.logo} src="/assets/bb-crest-34c67a3e6293d4779a6cc93c1dcec5d5724968d117976abb1651536a0835175d.png"></Image>
               <HeaderText>
                 <HeaderLogoColumn>
                   <HeaderLogoRow>